rework on the hooksystem, moved to new module
