+++ cms-api/src/main/java/com/condation/cms/api/hooks/HookSystem.java
@@ -20,14 +20,6 @@ package com.condation.cms.api.hooks;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-import com.condation.cms.api.annotations.Filter;
-import com.condation.cms.api.annotations.Action;
-import com.condation.cms.api.utils.AnnotationsUtil;
-import com.google.common.collect.ArrayListMultimap;
-import com.google.common.collect.Multimap;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 
@@ -37,78 +29,21 @@ import lombok.extern.slf4j.Slf4j;
  *
  * @author t.marx
  */
-@Slf4j
-public class HookSystem {
+public interface HookSystem {
 
-	Multimap<String, ActionHook> actions = ArrayListMultimap.create();
+	void register(Object sourceObject);
 
-	Multimap<String, FilterHook> filters = ArrayListMultimap.create();
+	public <T> void registerAction(final String name, final ActionFunction<T> hookFunction);
 
-	public HookSystem () {
-		
-	}
-	public HookSystem(HookSystem source) {
-		this.actions.putAll(source.actions);
-		this.filters.putAll(source.filters);
-	}
+	public <T> void registerAction(final String name, final ActionFunction<T> hookFunction, int priority);
 
-	public void register(Object sourceObject) {
-		// Action-Methoden registrieren
-		List<AnnotationsUtil.CMSAnnotation<Action, Void>> actionMethods
-				= AnnotationsUtil.process(sourceObject, Action.class, List.of(ActionContext.class), Void.class);
+	public <T> void registerFilter(final String name, final FilterFunction<T> hookFunction);
 
-		for (AnnotationsUtil.CMSAnnotation<Action, Void> ann : actionMethods) {
-			Action annotation = ann.annotation();
-			registerAction(annotation.value(), context -> ann.invoke(context), annotation.priority());
-		}
+	public <T> void registerFilter(final String name, final FilterFunction<T> hookFunction, int priority);
 
-		// Filter-Methoden registrieren
-		List<AnnotationsUtil.CMSAnnotation<Filter, Object>> filterMethods
-				= AnnotationsUtil.process(sourceObject, Filter.class, List.of(FilterContext.class), Object.class);
+	public ActionContext<Object> doAction(final String name);
 
-		for (AnnotationsUtil.CMSAnnotation<Filter, Object> ann : filterMethods) {
-			Filter annotation = ann.annotation();
-			registerFilter(annotation.value(), context -> ann.invoke(context), annotation.priority());
-		}
-	}
-
-	public <T> void registerAction(final String name, final ActionFunction<T> hookFunction) {
-		registerAction(name, hookFunction, 10);
-	}
-
-	public <T> void registerAction(final String name, final ActionFunction<T> hookFunction, int priority) {
-		actions.put(name, new ActionHook<>(name, priority, hookFunction));
-	}
-
-	public <T> void registerFilter(final String name, final FilterFunction<T> hookFunction) {
-		registerFilter(name, hookFunction, 10);
-	}
-
-	public <T> void registerFilter(final String name, final FilterFunction<T> hookFunction, int priority) {
-		filters.put(name, new FilterHook<>(name, priority, hookFunction));
-	}
-
-	public ActionContext<Object> doAction(final String name) {
-		return HookSystem.this.doAction(name, Map.of());
-	}
-
-	public ActionContext<Object> doAction(final String name, final Map<String, Object> arguments) {
-		var context = new ActionContext(new HashMap<>(arguments), new ArrayList<>());
-		actions.get(name).stream()
-				.sorted((h1, h2) -> Integer.compare(h1.priority(), h2.priority()))
-				.map((action) -> {
-					try {
-						return action.function().apply(context);
-					} catch (Exception e) {
-						log.error("error executing action", e);
-					}
-					return null;
-				})
-				.filter(value -> value != null)
-				.forEach(context.results()::add);
-
-		return context;
-	}
+	public ActionContext<Object> doAction(final String name, final Map<String, Object> arguments);
 
 	/**
 	 * calls all filters with the given parameters, if no filter is executed,
@@ -119,22 +54,5 @@ public class HookSystem {
 	 * @param parameters
 	 * @return
 	 */
-	public <T> FilterContext<T> doFilter(final String name, final T parameters) {
-		final FilterContext<T> returnContext = new FilterContext(
-				parameters
-		);
-		filters.get(name).stream()
-				.sorted((h1, h2) -> Integer.compare(h1.priority(), h2.priority()))
-				.forEach((var action) -> {
-					try {
-						var context = new FilterContext(returnContext.value());
-						var result = action.function().apply(context);
-						returnContext.value((T) result);
-					} catch (Exception e) {
-						log.error("error on filter", e);
-					}
-				});
-
-		return returnContext;
-	}
+	public <T> FilterContext<T> doFilter(final String name, final T parameters);
 }
