+++ cms-content/src/main/java/com/condation/cms/content/DefaultContentRenderer.java
@@ -50,8 +50,8 @@ import com.condation.cms.content.pipeline.ContentPipelineFactory;
 import com.condation.cms.content.views.model.View;
 import com.condation.cms.api.content.MapAccess;
 import com.condation.cms.content.pipeline.HTMLPipeline;
-import com.condation.cms.extensions.hooks.DBHooks;
-import com.condation.cms.extensions.hooks.TemplateHooks;
+import com.condation.cms.hooksystem.extensions.DBHooks;
+import com.condation.cms.hooksystem.extensions.TemplateHooks;
 import com.condation.cms.content.template.functions.LinkFunction;
 import com.condation.cms.content.template.functions.MarkdownFunction;
 import com.condation.cms.content.template.functions.list.NodeListFunctionBuilder;
