+++ cms-hooksystem/pom.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.condation.cms</groupId>
+        <artifactId>cms-parent</artifactId>
+        <version>8.1.0</version>
+    </parent>
+    <artifactId>cms-hooksystem</artifactId>
+    <packaging>jar</packaging>
+	<name>CMS HookSystem</name>
+    
+	<dependencies>
+		<dependency>
+			<groupId>com.condation.cms</groupId>
+			<artifactId>cms-api</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.condation.cms</groupId>
+			<artifactId>cms-extensions</artifactId>
+		</dependency>
+		
+		<dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <scope>provided</scope>
+        </dependency>
+	</dependencies>
+</project>
