+++ cms-hooksystem/src/main/java/com/condation/cms/hooksystem/ActionHook.java
@@ -1,4 +1,6 @@
-package com.condation.cms.api.hooks;
+package com.condation.cms.hooksystem;
+
+import com.condation.cms.api.hooks.ActionFunction;
 
 /*-
  * #%L
