+++ cms-hooksystem/src/main/java/com/condation/cms/hooksystem/extensions/TemplateComponentsWrapper.java
@@ -1,4 +1,4 @@
-package com.condation.cms.extensions.hooks;
+package com.condation.cms.hooksystem.extensions;
 
 /*-
  * #%L
