+++ cms-hooksystem/src/test/resources/site/extensions/hook-test.js
