+++ cms-server/src/main/java/com/condation/cms/request/RequestContextFactory.java
@@ -35,7 +35,6 @@ import com.condation.cms.api.feature.features.HookSystemFeature;
 import com.condation.cms.api.feature.features.InjectorFeature;
 import com.condation.cms.api.feature.features.IsDevModeFeature;
 import com.condation.cms.api.feature.features.MarkdownRendererFeature;
-import com.condation.cms.api.feature.features.ModuleManagerFeature;
 import com.condation.cms.api.feature.features.RequestFeature;
 import com.condation.cms.api.feature.features.ServerPropertiesFeature;
 import com.condation.cms.api.feature.features.SiteMediaServiceFeature;
@@ -56,11 +55,11 @@ import com.condation.cms.content.RenderContext;
 import com.condation.cms.content.tags.Tags;
 import com.condation.cms.content.tags.TagParser;
 import com.condation.cms.extensions.ExtensionManager;
-import com.condation.cms.extensions.hooks.ContentHooks;
-import com.condation.cms.extensions.hooks.DBHooks;
-import com.condation.cms.extensions.hooks.ServerHooks;
-import com.condation.cms.extensions.hooks.TemplateHooks;
 import com.condation.cms.extensions.request.RequestExtensions;
+import com.condation.cms.hooksystem.extensions.ContentHooks;
+import com.condation.cms.hooksystem.extensions.DBHooks;
+import com.condation.cms.hooksystem.extensions.ServerHooks;
+import com.condation.cms.hooksystem.extensions.TemplateHooks;
 import com.condation.modules.api.ModuleManager;
 import com.google.inject.Injector;
 import java.io.IOException;
