+++ cms-server/src/main/java/com/condation/cms/server/handler/http/APIHandler.java
@@ -30,9 +30,8 @@ import com.condation.cms.api.feature.features.ConfigurationFeature;
 import com.condation.cms.api.request.RequestContext;
 import com.condation.cms.api.utils.RequestUtil;
 import com.condation.cms.extensions.HttpHandlerExtension;
-import com.condation.cms.extensions.hooks.ServerHooks;
 import com.condation.cms.extensions.http.JettyHttpHandlerWrapper;
-import com.condation.cms.server.filter.CreateRequestContextFilter;
+import com.condation.cms.hooksystem.extensions.ServerHooks;
 import com.condation.cms.server.handler.AbstractHandler;
 import com.condation.modules.api.ModuleManager;
 import com.google.inject.Inject;
@@ -40,7 +39,6 @@ import com.google.inject.Inject;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.Response;
 import org.eclipse.jetty.util.Callback;
