+++ cms-server/src/main/java/com/condation/cms/server/handler/http/RoutesHandler.java
@@ -29,12 +29,10 @@ import com.condation.cms.api.extensions.http.routes.RoutesManager;
 import com.condation.cms.api.request.RequestContext;
 import com.condation.cms.api.utils.RequestUtil;
 import com.condation.cms.extensions.HttpHandlerExtension;
-import com.condation.cms.extensions.hooks.ServerHooks;
 import com.condation.cms.extensions.http.JettyHttpHandlerWrapper;
-import com.condation.cms.server.filter.CreateRequestContextFilter;
+import com.condation.cms.hooksystem.extensions.ServerHooks;
 import com.condation.modules.api.ModuleManager;
 import com.google.inject.Inject;
-import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
