+++ cms-server/src/main/java/com/condation/cms/server/host/VHost.java
@@ -53,8 +53,8 @@ import com.condation.cms.api.theme.Theme;
 import com.condation.cms.core.utils.SiteUtil;
 import com.condation.cms.core.configuration.ConfigManagement;
 import com.condation.cms.extensions.GlobalExtensions;
-import com.condation.cms.extensions.hooks.GlobalHooks;
 import com.condation.cms.filesystem.FileDB;
+import com.condation.cms.hooksystem.extensions.GlobalHooks;
 import com.condation.cms.media.MediaManager;
 import com.condation.cms.media.SiteMediaManager;
 import com.condation.cms.media.ThemeMediaManager;
