+++ cms-templates/pom.xml
@@ -36,6 +36,11 @@
 		</dependency>
 		
 		<dependency>
+			<groupId>com.condation.cms</groupId>
+			<artifactId>cms-hooksystem</artifactId>
+		</dependency>
+		
+		<dependency>
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-jexl3</artifactId>
 		</dependency>
