+++ cms-templates/src/test/java/com/condation/cms/templates/TemplateEngineFunctionsTest.java
@@ -29,7 +29,8 @@ import com.condation.cms.api.feature.features.ModuleManagerFeature;
 import com.condation.cms.api.hooks.HookSystem;
 import com.condation.cms.api.model.Parameter;
 import com.condation.cms.api.request.RequestContext;
-import com.condation.cms.extensions.hooks.TemplateHooks;
+import com.condation.cms.hooksystem.CMSHookSystem;
+import com.condation.cms.hooksystem.extensions.TemplateHooks;
 import com.condation.cms.templates.components.TemplateComponents;
 import com.condation.cms.templates.loaders.StringTemplateLoader;
 import com.condation.modules.api.ModuleManager;
@@ -71,7 +72,7 @@ public class TemplateEngineFunctionsTest extends AbstractTemplateEngineTest {
 	@Test
 	void getFunctionsFromDynamicConig () {
 		var requestContext = new RequestContext();
-		requestContext.add(HookSystemFeature.class, new HookSystemFeature(new HookSystem()));
+		requestContext.add(HookSystemFeature.class, new HookSystemFeature(new CMSHookSystem()));
 		requestContext.add(TemplateHooks.class, new TemplateHooks(requestContext));
 		
 		var injectorMock = Mockito.mock(Injector.class);
