+++ integration-tests/pom.xml
@@ -47,6 +47,10 @@
 		</dependency>
 		<dependency>
 			<groupId>com.condation.cms</groupId>
+			<artifactId>cms-hooksystem</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>com.condation.cms</groupId>
 			<artifactId>cms-media</artifactId>
 		</dependency>
 		<dependency>
