+++ modules/ui-module/pom.xml
@@ -24,6 +24,10 @@
 			<artifactId>cms-auth</artifactId>
 		</dependency>
 		<dependency>
+			<groupId>com.condation.cms</groupId>
+			<artifactId>cms-hooksystem</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>io.jsonwebtoken</groupId>
 			<artifactId>jjwt-api</artifactId>
 			<version>0.13.0</version>
