+++ pom.xml
@@ -30,6 +30,7 @@
         <module>integration-tests</module>
         <module>cms-core</module>
         <module>cms-test</module>
+        <module>cms-hooksystem</module>
     </modules>
     <inceptionYear>2023</inceptionYear>
     <organization>
@@ -106,6 +107,11 @@
             <dependency>
                 <groupId>com.condation.cms</groupId>
                 <artifactId>cms-test</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+			<dependency>
+                <groupId>com.condation.cms</groupId>
+                <artifactId>cms-hooksystem</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
