+++ test-server/modules/example-module/module.properties
@@ -1,4 +1,4 @@
 id=example-module
 name=example module
-version=8.0.0
+version=8.1.0
 priority=HIGH
